Move Run Tab components into new RunSection
